requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
passlib[argon2]==1.7.4
sqlalchemy==2.0.23
python-dotenv==1.0.1
pydantic==1.10.12
python-multipart==0.0.6
alembic==1.12.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2